Add an option to disable building documentation
This is useful for private builds from git as c2man is not widely
available, those builds might not need documentation and unlike in
tarballs there's no prebuilt docs in git.

This was already attempted previously in 21b18b07c0b57fbaff15b622df80408cf8936809
but then reverted in 095d88503194e81059ab81d04b849d9245899dfa as
the approach taken resulted in configure requiring c2man even for builds
from tarballs unless opting out of installing the already built docs.

The approach taken this time changes nothing when --disable-docs
is not passed, so there's no superfluous c2man requirement.
If --disable-docs is passed, documentation will not be built or
installed. When attempting to create a distribution tarball
from a setup with --disable-docs it will error out early as the
buildsystem still knows the tarball should include the docs but
doesn't know how to build them.
This avoids accidentally building a tarball without prebuilt docs.
(If the conditional around the targets were omitted, the rules would
 be invoked regardless, generating the documentation if c2man is
 available. But later on distcheck would still error out because after
 distclean the newly built documentation still exists. The error
 message might be less clear though.)

An alternative and slightly more concise approach would be to
conditionally omit the doc SUBDIR, but this might (I haven't checked)
affect MAINTAINERCLEANFILES and derived targets in git.mk.

Yet another alternative would be to adopt the original
21b18b07c0b57fbaff15b622df80408cf8936809 with an additional
check probing for the existinece of one prebuilt manpage similar
to what the Meson build does.
This would be less concise though and checking only for one prebuilt
manpage might theoretically cause issues in some odd cases.

Fixes: https://github.com/fribidi/fribidi/issues/8 (again)